refactored to be in package com.tacocardgame.model
