--- src/DeleteMe.java
@@ -1,2 +1,0 @@
-class DeleteMe {
-}
