+++ src/com/tacocardgame/model/player.java
@@ -1,4 +1,4 @@
-package com.tcgcp.player;
+package com.tacocardgame.model;
 
 import java.util.ArrayList;
