+++ src/com/tacocardgame/model/user.java
@@ -1,4 +1,4 @@
-package com.tcgcp.player;
+package com.tacocardgame.model;
 
 public class user extends player {
